refactor(llmobs): drop InjectContext/ExtractContext, consolidate tests
Defer distributed agent attribution propagation to a follow-up PR.
The in-process attribution (resolveParentAgent + meta.agent_attribution
serialization) is self-contained and ships without the distributed API.

- Delete propagation.go and its tests (InjectContext, ExtractContext,
  injectAgentAttribution, extractAgentAttribution, agentNameWireSafe)
- Remove InjectContext/ExtractContext from the public llmobs package
- Merge resolve_parent_agent_test.go into start_span_agent_test.go
